pbar fix

+++ src/cdl_eeg/models/main_models/main_fixed_channels_model.py
@@ -300,7 +300,7 @@ class MainFixedChannelsModel(nn.Module):
         best_model_state = {k: v.cpu() for k, v in self.state_dict().items()}
         for epoch in range(num_epochs):
             # Start progress bar
-            pbar = enlighten.Counter(total=numpy.ceil(len(train_loader.dataset) / train_loader.batch_size),
+            pbar = enlighten.Counter(total=int(numpy.ceil(len(train_loader.dataset) / train_loader.batch_size)),
                                      desc=f"Epoch {epoch + 1}/{num_epochs}", unit="batch")
 
             # ---------------
